Create useRooom Hook to cleanup Room component
